all passing tests, delete of scroll if no file
