+++ Cargo.toml
@@ -3,6 +3,10 @@ name = "legatio"
 version = "0.1.3"
 edition = "2021"
 
+[lib]
+name = "legatio"
+path = "src/lib.rs"
+
 [profile.dev]
 debug = 1
