+++ src/lib.rs
@@ -1,0 +1,20 @@
+pub mod core;
+pub mod utils;
+pub mod services;
+
+pub use crate::core::prompt::*;
+pub use crate::core::project::*;
+pub use crate::core::scroll::*;
+pub use crate::core::canvas::*;
+
+pub use crate::utils::structs::*;
+pub use crate::utils::db_utils::*;
+pub use crate::utils::error::*;
+pub use crate::utils::logger::*;
+
+pub use crate::services::legatio::*;
+pub use crate::services::display::*;
+pub use crate::services::model::*;
+pub use crate::services::search::*;
+pub use crate::services::ui::*;
+
