+++ src/utils/db_utils.rs
@@ -119,173 +119,3 @@ pub async fn delete_module(
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use sqlx::SqlitePool;
-
-    // A test database URL (using an in-memory SQLite database for isolated testing)
-    const TEST_DATABASE_URL: &str = "sqlite::memory:";
-
-    // Helper function to set up the database
-    async fn setup_test_db() -> SqlitePool {
-        // Call the function to create the database and tables
-
-        get_db_pool(TEST_DATABASE_URL)
-            .await
-            .expect("Failed to set up test database")
-    }
-
-    #[tokio::test]
-    async fn test_get_db_pool_creates_tables() {
-        let pool = setup_test_db().await;
-
-        // Here, we'll check if the tables actually exist in the in-memory database
-        let result_projects: Result<(String,), _> =
-            sqlx::query_as("SELECT name FROM sqlite_master WHERE type='TABLE' AND name='projects'")
-                .fetch_one(&pool)
-                .await;
-
-        let result_scrolls: Result<(String,), _> =
-            sqlx::query_as("SELECT name FROM sqlite_master WHERE type='table' AND name='scrolls'")
-                .fetch_one(&pool)
-                .await;
-
-        let result_prompts: Result<(String,), _> =
-            sqlx::query_as("SELECT name FROM sqlite_master WHERE type='table' AND name='prompts'")
-                .fetch_one(&pool)
-                .await;
-
-        assert!(
-            result_projects.is_ok(),
-            "Projects table was not created as expected: \n{:?}",
-            result_projects
-        );
-        assert!(
-            result_scrolls.is_ok(),
-            "Scrolls table was not created as expected"
-        );
-        assert!(
-            result_prompts.is_ok(),
-            "Prompts table was not created as expected"
-        );
-    }
-
-    //#[tokio::test]
-    //async fn test_delete_module_success() {
-    //    let pool = setup_test_db().await;
-
-    //    // Insert a dummy record into the projects table
-    //    let project_id = "test_project_id";
-    //    sqlx::query("INSERT INTO projects (project_id, project_path) VALUES (?, ?)")
-    //        .bind(project_id)
-    //        .bind("/test/path")
-    //        .execute(&pool)
-    //        .await
-    //        .expect("Failed to insert dummy project");
-
-    //    // Verify the record exists
-    //    let project_exists = sqlx::query_as::<_>("SELECT COUNT(*) FROM projects WHERE project_id = ?")
-    //        .bind(project_id)
-    //        .fetch_one(&pool)
-    //        .await
-    //        .unwrap()
-
-    //    assert_eq!(project_exists, 1, "Dummy project record was not added");
-
-    //    // Delete the record using the function
-    //    delete_module(&pool, "projects", "project_id", project_id)
-    //        .await
-    //        .expect("Failed to delete project");
-
-    //    // Verify the record no longer exists
-    //    let project_exists = sqlx::query("SELECT COUNT(*) FROM projects WHERE project_id = ?")
-    //        .bind(project_id)
-    //        .fetch_one(&pool)
-    //        .await
-    //        .expect("Failed to verify project non-existence")
-    //        .get::<i64, _>("COUNT(*)");
-
-    //    assert_eq!(project_exists, 0, "Project record was not successfully deleted");
-    //}
-
-    //#[tokio::test]
-    //async fn test_delete_module_non_existent_record() {
-    //    let pool = setup_test_db().await;
-
-    //    let result = delete_module(&pool, "projects", "project_id", "non_existent_id").await;
-
-    //    // Expect no error as the function should handle the "no record found" case gracefully
-    //    assert!(result.is_ok(), "delete_module should not fail when deleting a non-existent record");
-    //}
-
-    #[tokio::test]
-    async fn test_delete_module_errors_on_invalid_table() {
-        let pool = setup_test_db().await;
-
-        // Attempt to delete from an invalid table name
-        let result = delete_module(&pool, "non_existent_table", "column_name", "value").await;
-
-        // Expect an error since the table doesn't exist
-        assert!(
-            result.is_err(),
-            "delete_module should fail when the table name is invalid"
-        );
-
-        // Verify the specific error type
-        if let Err(AppError::DatabaseError(message)) = result {
-            assert!(
-                message.contains("no such table"),
-                "Error message does not mention 'no such table'"
-            );
-        } else {
-            panic!("Expected AppError::DatabaseError, but received a different error.");
-        }
-    }
-
-    #[tokio::test]
-    async fn test_get_db_pool_handles_reconnection() {
-        // Test reconnection to an already-existing database
-        {
-            let first_pool = get_db_pool(TEST_DATABASE_URL)
-                .await
-                .expect("Failed to create first database connection");
-            assert!(!first_pool.is_closed(), "First pool is unexpectedly closed");
-        }
-
-        // Simulate reconnecting
-        {
-            let second_pool = get_db_pool(TEST_DATABASE_URL)
-                .await
-                .expect("Failed to reconnect to the database");
-            assert!(
-                !second_pool.is_closed(),
-                "Second pool is unexpectedly closed"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_db_error_handling_on_failed_connection() {
-        // Provide an invalid database URL to simulate a failure
-        let invalid_db_url = "sqlite:/invalid_path";
-
-        let result = get_db_pool(invalid_db_url).await;
-
-        // Expect an error
-        assert!(
-            result.is_err(),
-            "get_db_pool should fail when provided with an invalid database URL"
-        );
-
-        // Verify the specific error type
-        if let Err(AppError::DatabaseError(message)) = result {
-            assert!(
-                message.contains("Failed to connect"),
-                "Error message does not mention 'Failed to connect'"
-            );
-        } else {
-            panic!("Expected AppError::DatabaseError, but received a different error.");
-        }
-    }
-}
