+++ src/utils/error.rs
@@ -1,20 +1,29 @@
-use thiserror::Error;
+use std::error::Error;
+use std::fmt;
 
-/// Define crate-wide error types using `thiserror`.
-#[derive(Error, Debug)]
+/// Define crate-wide error types without using `thiserror`.
+#[derive(Debug)]
 pub enum AppError {
-    #[error("Database operation failed: {0}")]
     DatabaseError(String),
-
-    #[error("File operation failed: {0}")]
     FileError(String),
-
-    #[error("Validation error: {0}")]
     ValidationError(String),
-
-    #[error("Unexpected or unknown error: {0}")]
     UnexpectedError(String),
 }
+
+/// Implement `std::fmt::Display` for `AppError`.
+impl fmt::Display for AppError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            AppError::DatabaseError(msg) => write!(f, "Database operation failed: {}", msg),
+            AppError::FileError(msg) => write!(f, "File operation failed: {}", msg),
+            AppError::ValidationError(msg) => write!(f, "Validation error: {}", msg),
+            AppError::UnexpectedError(msg) => write!(f, "Unexpected or unknown error: {}", msg),
+        }
+    }
+}
+
+/// Implement `std::error::Error` for `AppError`.
+impl Error for AppError {}
 
 /// Custom Result type that uses `AppError`.
 pub type Result<T, E = AppError> = std::result::Result<T, E>;
